requirements.txt
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
pgvector>=0.2.4
python-multipart>=0.0.6
python-docx>=1.1.0
PyPDF2>=3.0.1
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.6.0
mlflow>=2.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
tenacity>=8.2.0
